Slimboy-Bot-2.2.4/requirements.txt

# Discord Bot Dependencies
discord.py>=2.5.2

# Web Server for Keep Alive
Flask>=3.1.1

# Additional utilities
psutil>=5.9.0
aiohttp>=3.8.0
discord.py
